+++ src/scolta/index/__init__.py
@@ -1,0 +1,1 @@
+"""Pagefind in-process indexer subsystem."""